@@ -1,23 +1,23 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <Description>The CQRS Queries library</Description>
-        <Authors>Konrad Dąbrowski</Authors>
-        <TargetFramework>net5.0</TargetFramework>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <LangVersion>Latest</LangVersion>
-        <AssemblyName>Arcaim.CQRS.Queries</AssemblyName>
-        <Title>Arcaim.CQRS.Queries</Title>
-        <PackageId>Arcaim.CQRS.Queries</PackageId>
-        <Company>Arcaim</Company>
-        <Version>1.3.0</Version>
-    </PropertyGroup>
+  <PropertyGroup>
+    <Description>The CQRS Queries library</Description>
+    <Authors>Konrad Dąbrowski</Authors>
+    <TargetFramework>net6.0</TargetFramework>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <LangVersion>Latest</LangVersion>
+    <AssemblyName>Arcaim.CQRS.Queries</AssemblyName>
+    <Title>Arcaim.CQRS.Queries</Title>
+    <PackageId>Arcaim.CQRS.Queries</PackageId>
+    <Company>Arcaim</Company>
+    <Version>6.0.0</Version>
+  </PropertyGroup>
 
-    <ItemGroup>
-        <PackageReference Include="Arcaim.DI.Scanner" Version="1.0.0" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.0">
-            <PrivateAssets>all</PrivateAssets>
-        </PackageReference>
-    </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Arcaim.DI.Scanner" Version="6.0.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0">
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
+  </ItemGroup>
 
 </Project>
