@@ -1,9 +1,8 @@
 using System.Threading.Tasks;
 
-namespace Arcaim.CQRS.Queries
+namespace Arcaim.CQRS.Queries;
+
+public interface IQueryHandler<in TQuery, TResult> where TQuery : class, IQuery<TResult>
 {
-    public interface IQueryHandler<in TQuery, TResult> where TQuery : class, IQuery<TResult>
-    {
-        Task<TResult> HandleAsync(TQuery query);
-    }
+  Task<TResult> HandleAsync(TQuery query);
 }
